any sin

+++ database/migrations/2024_07_10_133838_add_columns_to_users_table.php
@@ -6,25 +6,28 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('phone', 255)->after('password');
-            $table->string('role')->default('user')->after('phone');
+            if (!Schema::hasColumn('users', 'phone')) {
+                $table->string('phone', 255)->after('password');
+            }
+            if (!Schema::hasColumn('users', 'role')) {
+                $table->string('role')->default('user')->after('phone');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('phone');
-            $table->dropColumn('role');
+            if (Schema::hasColumn('users', 'phone')) {
+                $table->dropColumn('phone');
+            }
+            if (Schema::hasColumn('users', 'role')) {
+                $table->dropColumn('role');
+            }
         });
     }
 };
+
