+++ database/migrations/2024_08_08_122240_add_location_to_users_table.php
@@ -6,23 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('location')->nullable()->after('profile_photo_path');
+            if (!Schema::hasColumn('users', 'location')) {
+                $table->string('location')->nullable()->after('profile_photo_path');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('location');
+            if (Schema::hasColumn('users', 'location')) {
+                $table->dropColumn('location');
+            }
         });
     }
 };
